Move newIngesterErrXXX tests to errors_test.go
